Simplify refdata import view to one file
